fix: input circularity
